requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
